+++ terminal/terminalapi/terminalapi.go
@@ -49,4 +49,8 @@ type Terminal interface {
 	// This call blocks until the next event or cancellation of the context.
 	// Returns nil when the context gets canceled.
 	Event(ctx context.Context) Event
+
+	// Close closes the underlying terminal implementation and should be called when
+	// the terminal isn't required anymore to return the screen to a sane state.
+	Close()
 }