+++ CHANGELOG.md
@@ -7,6 +7,26 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.11.0] - 7-Mar-2020
+
+#### Breaking API changes
+
+- Termdash now requires at least Go version 1.11.
+
+### Added
+
+- New [`tcell`](https://github.com/gdamore/tcell) based terminal implementation
+  which implements the `terminalapi.Terminal` interface.
+- tcell implementation supports two initialization `Option`s:
+  - `ColorMode` the terminal color output mode (defaults to 256 color mode)
+  - `ClearStyle` the foreground and background color style to use when clearing
+     the screen (defaults to the global ColorDefault for both foreground and
+     background)
+
+### Fixed
+
+- Improved test coverage of the `Gauge` widget.
+
 ## [0.10.0] - 5-Jun-2019
 
 ### Added
@@ -295,7 +315,8 @@ identifiers shouldn't be used externally.
 - The Gauge widget.
 - The Text widget.
 
-[unreleased]: https://github.com/mum4k/termdash/compare/v0.10.0...devel
+[unreleased]: https://github.com/mum4k/termdash/compare/v0.11.0...devel
+[0.11.0]: https://github.com/mum4k/termdash/compare/v0.10.0...v0.11.0
 [0.10.0]: https://github.com/mum4k/termdash/compare/v0.9.1...v0.10.0
 [0.9.1]: https://github.com/mum4k/termdash/compare/v0.9.0...v0.9.1
 [0.9.0]: https://github.com/mum4k/termdash/compare/v0.8.0...v0.9.0